requirement.txt
kafka-python
dotenv
pyspark
findspark
json
tweepy
